backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.6.0
sqlalchemy>=2.0.0
werkzeug>=3.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
email-validator>=2.1.0.post1
